move custom component to dom rendering
